Feat/85/delete groupe authorization

+++ src/modules/groups/application/services/groups.service.ts
@@ -1,14 +1,13 @@
 import { EntityRepository } from '@mikro-orm/mysql';
 import { InjectRepository } from '@mikro-orm/nestjs';
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { randomUUID } from 'crypto';
-import {
-  Groups,
-} from 'src/modules/groups/domain/entities/groups.entity';
+import { Groups } from 'src/modules/groups/domain/entities/groups.entity';
 import { UsersService } from 'src/modules/users/application/services/users.service';
 import { CreateGroupDto } from '../dto/create-group.dto';
 import { ModifyGroupDto } from '../dto/modify-group.dto';
 import { User } from '../../../users/domain/entities/user.entity';
+import { DeleteGroupDto } from '../dto/delete-group.dto';
 
 @Injectable()
 export class GroupsService {
@@ -46,9 +45,18 @@ export class GroupsService {
     await this.groupsRepository.flush();
   }
 
-  async remove(groupUuid: string): Promise<void> {
-    // TODO : autorisation
-    const group = await this.groupsRepository.findOne(groupUuid);
+  async remove(deleteGroupDto: DeleteGroupDto): Promise<void> {
+    const user: User = await this.userService.findOneByUuid(deleteGroupDto.userUuid);
+
+    if(!user){
+      throw new BadRequestException("User unknown");
+    }
+    const group = await this.groupsRepository.findOne(deleteGroupDto.groupUuid);
+
+    if(group.getCreator() !== user) {
+      throw new ForbiddenException("Operation denied");
+    }
+
     await this.groupsRepository.removeAndFlush(group);
   }
 
@@ -61,8 +69,13 @@ export class GroupsService {
   }
 
   async getUserGroups(userUuid: string): Promise<Array<Groups>> {
-    const user: User = await this.userService.findOneByUuid(userUuid)
+    const user: User = await this.userService.findOneByUuid(userUuid);
 
-    return await this.groupsRepository.find({createdBy: user})
+    if(!user){
+      throw new BadRequestException("User unknown");
+    }
+
+    // @ts-ignore
+    return this.groupsRepository.find({ createdBy: user });
   }
 }
